+++ src/main/java/org/springframework/data/jpa/util/Jpa21Utils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.jpa.repository.query;
+package org.springframework.data.jpa.util;
 
 import java.lang.reflect.Method;
 
@@ -21,12 +21,13 @@ import javax.persistence.EntityGraph;
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
 
+import org.springframework.data.jpa.repository.query.JpaEntityGraph;
 import org.springframework.util.Assert;
 import org.springframework.util.ClassUtils;
 import org.springframework.util.ReflectionUtils;
 
 /**
- * Utils for brideing various JPA 2.1 features.
+ * Utils for bridging various JPA 2.1 features.
  * 
  * @author Thomas Darimont
  * @since 1.6
@@ -72,18 +73,19 @@ public enum Jpa21Utils {
 	 * 
 	 * @see JPA 2.1 Specfication 3.7.4 - Use of Entity Graphs in find and query operations P.117
 	 * @param em must not be {@literal null}
-	 * @param entityGraph must not be {@literal null}
+	 * @param jpaEntityGraph must not be {@literal null}
+	 * @return the {@link EntityGraph} described by the given {@code entityGraph} 
 	 */
-	public EntityGraph<?> tryGetFetchGraph(EntityManager em, JpaEntityGraph entityGraph) {
+	public EntityGraph<?> tryGetFetchGraph(EntityManager em, JpaEntityGraph jpaEntityGraph) {
 
 		Assert.notNull(em, "EntityManager must not be null!");
-		Assert.notNull(entityGraph, "EntityGraph must not be null!");
+		Assert.notNull(jpaEntityGraph, "EntityGraph must not be null!");
 
 		Assert.isTrue(JPA21_AVAILABLE, "The EntityGraph-Feature requires at least a JPA 2.1 persistence provider!");
 		Assert.isTrue(GET_ENTITY_GRAPH_METHOD != null,
 				"It seems that you have the JPA 2.1 API but a JPA 2.0 implementation on the classpath!");
 
-		EntityGraph<?> graph = em.getEntityGraph(entityGraph.getName());
+		EntityGraph<?> graph = em.getEntityGraph(jpaEntityGraph.getName());
 
 		return graph;
 	}
